requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
coverage==7.8.2
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pluggy==1.6.0
-e git+https://github.com/code72139/project_AD.git@380186fad6ccfaccc1aaef71f25d35739af1ac07#egg=project_ad
psycopg2==2.9.10
pytest==8.3.5
pytest-cov==6.1.1
pytest-flask==1.3.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
